requirements.txt
torch
torchvision
numpy
pandas
opencv-python
pillow
matplotlib
seaborn
scikit-learn
streamlit
tensorboard
kaggle
scikit-image
lpips
